refactor workflowstatustype
